+++ ibllib/pipes/ephys_preprocessing.py
@@ -56,7 +56,7 @@ class EphysAudio(tasks.Task):
     level = 0  # this job doesn't depend on anything
 
     def _run(self, overwrite=False):
-        command = 'ffmpeg -i {file_in} -y -c:a flac -nostats {file_out}'
+        command = 'ffmpeg -i {file_in} -y -nostdin -c:a flac -nostats {file_out}'
         file_in = next(self.session_path.rglob('_iblrig_micData.raw.wav'), None)
         if file_in is None:
             return
@@ -155,8 +155,8 @@ class EphysVideoCompress(tasks.Task):
 
     def _run(self, **kwargs):
         # avi to mp4 compression
-        command = ('ffmpeg -i {file_in} -y -codec:v libx264 -preset slow -crf 17 '
-                   '-nostats -loglevel 0 -codec:a copy {file_out}')
+        command = ('ffmpeg -i {file_in} -y -nostdin -codec:v libx264 -preset slow -crf 17 '
+                   '-loglevel 0 -codec:a copy {file_out}')
         output_files = ffmpeg.iblrig_video_compression(self.session_path, command)
         if len(output_files) == 0:
             self.session_path.joinpath('')
@@ -228,9 +228,11 @@ class EphysMtscomp(tasks.Task):
                 sr = spikeglx.Reader(bin_file)
                 if sr.is_mtscomp:
                     out_files.append(bin_file)
+                    out_files.append(bin_file.with_suffix('.ch'))
                 else:
                     _logger.info(f"Compressing binary file {bin_file}")
                     out_files.append(sr.compress_file(keep_original=False))
+                    out_files.append(bin_file.with_suffix('.ch'))
         return out_files